BackDevelop/requirements.txt
Django>=4.2.4
django-admin-datta>=1.0.10
django-cors-headers>=4.2.0
django-webpack-loader>=2.0.1
djangorestframework>=3.14.0
mysqlclient >= 2.2.0
djangorestframework-simplejwt >= 5.3.0
django-rest-swagger >= 2.2.0
drf-yasg>=1.21.7
django-filter>=23.3
django-s3-storage>=0.14.0
Pillow>=10.0.1
django-rest-passwordreset>=1.3.0

uvicorn[standard]
websockets
wsproto
channels-redis

# run : uvicorn core.asgi:application --port 8001
channels~=4.0.0